Created kickstart.py for hopla startup code
