+++ setup.cfg
@@ -51,4 +51,4 @@ exclude = hopla.tests
 [options.entry_points]
 # syntax: short_name = package.module:function
 console_scripts =
-    hopla = hopla:entry_cmd
+    hopla = hopla:setup_hopla_cli
